Implement build process with gulp
I want react-router for routing and to do that, I'll need to require it
using browserify. To use browserify cleanly, I'll need a build process. I
also need a build process to transpile the JSX to JS so I've set up a
preliminary one using gulp (no transpiling ATM but it does concatenate and
copy the JSX files)