Consistent-ize props in Common and SocketView
I explicitly mentioned the value and onChange props in the common
components. In SocketView I moved the creaiton of the bff into the default
props step, just like the homeview. Also linked changes in the inner
toggle & range to the change functions in the view, which hit the bff lib
and will, in the future, hit the backend.