Edit HomeView to load items after mount and retain items as state.
There was an issue doing the loading inside the actual render call -
promises are async and it wouldn't load the actual items before the
rendering was complemeted. By moving items into state and doing it after
the mount, I ensured the screen would show as soon as possible and the
items would be loaded.