Edit TabItem to make use of activeClassName
Link has a built-in activeClassName for checking whether or not you are
actual on the url the link points to, so I made use of it. I also needed
to include a conditional for the cases when you want to navigate to '/',
which would usually mean that would also be active, in which case you want
to use IndexLink.