Rename HomeView to SocketListView
For now, I've opted to implement Tabs as a separate page, which keeps
things simple, though pretty much the same code is duplicated. When I
implement transitions, this structure is going to make things werid
because it's going to transition the header too when all I want is the
Content to transition. Perhaps I should put a header into <App />...?
Well, I could do that. I'd just need to move the login and register routes
outside...