Make it so that only one SocketItem in a SocketList can be selected @once
After browsing through the React docs once more, I figured out how to only
have one SocketItem toggled (revealing the range) at one time. It's very
similar to a radio-button setup and needs to be manage by the parent
element, so that's what I did. By setting up the touch handler in the
child list elements but calling a function on the parent, I retain an
overlooking scope while still allowing for complex multitouch event
handlers to be established. Also this pattern can be replicated with most
of the other handlers too, which opens an avenue for putting more focus
(and function) on the SocketList rather than the SocketItem.